contiuues to put some mutation refactor

+++ src/resolvers/Mutation.js
@@ -44,24 +44,25 @@ const Mutation = {
             token: jwt.sign({ userId: user.id }, 'thisisasecret')
         }
     },
-    async deleteUser(parent, args, { prisma }, info) {
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.deleteUser({ 
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.updateUser({ 
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
     },
     createPost(parent, args, { prisma, request }, info) {
         const userId = getUserId(request)
-
         return prisma.mutation.createPost({ 
             data: {
                 title: args.data.title,
@@ -75,14 +76,35 @@ const Mutation = {
             }
         }, info)
     },
-    deletePost(parent, args, { prisma, pubsub }, info) {
+     async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        }) 
+        if (!postExists) {
+            throw new Error('Cannot delete post')
+        }
         return prisma.mutation.deletePost({ 
             where: {
                 id: args.id
             }
         }, info)
     },
-    updatePost(parent, args, { prisma, subsub }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const postBelongsToUser = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        }) 
+        if (!postBelongsToUser) {
+            throw new Error('Post does not belong to you')
+        }
         return prisma.mutation.updatePost({ 
             where: {
                 id: args.id
@@ -90,13 +112,14 @@ const Mutation = {
             data: args.data
         }, info)
     },
-    createComment(parent, args, { prisma, pubsub }, info) {
+    createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.createComment({ 
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -107,17 +130,34 @@ const Mutation = {
             }
         }, info)
     },
-    deleteComment(parent, args, { prisma, pubsub }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentBelongsToUser = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if (!commentBelongsToUser) { throw new Error('Cant delete comment') }
+
         return prisma.mutation.deleteComment({ 
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma, pubsub }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if (!commentExists) { throw new Error('Not your comment')}
         return prisma.mutation.updateComment({ 
             where: {
-                id: args.id
+                id: args.id,
             },
             data: args.data
         }, info)
